requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
email_validator
WTForms-SQLAlchemy
Gunicorn
pytest
pytest-flask
pytest-cov
